Rename multiconnectionmanager module to connectionmanager.
There's no longer any singleconnectionmanager so the distinction is
unnecessary.
