+++ src/device/socket/connectionmanager.rs
